Break out component helpers into their own module
